+++ sdk/eventhub/event-hubs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 5.2.2 (Unreleased)
 
+- Fixes issue [#9289](https://github.com/Azure/azure-sdk-for-js/issues/9289)
+  where calling `await subscription.close()` inside of a subscription's `processError`
+  handler would cause the subscription to deadlock.
 
 ## 5.2.1 (2020-06-08)